Refactor restructure into atomic operations
The restructure plugin was initially created where each operation type
had an implicit order. This refactor changes it so now a user specifies
an array of operations, their types, and their parameters, and those are
executed in order.

Additionally, it removes the FieldSelector interface and instead makes
it just a concrete type. The interface was a layer of indirection more
than necessary.