Add details to error message (#55)
When external packages import the log agent (the UA), they don't use the
errors package or zap, so the additional metadata on the error messages
is lost. Since we depend on that data to include useful debugging
information, we should add the details to the error message. This commit
marshals the details as JSON if they exist and appends them to the
description.