Improve stability of file input plugin
Previously, if a file was truncated between the time that we read a line
and we try to save the fingerprint, it threw an error because we would
get an unexpected EOF when reading n bytes, because n is calculated
before reading.

Now, we simply read up to n (our offset) bytes. This was causing some
test stability issues that would occasionally cause build failures.