Fix reading partial logs in active files
When logs are actively being written to, we can't assume that an EOF
is a valid log entry terminator since the application writing could
have written a partial entry.

This commit changes our logic so we only read to the end of the file
if both the offset and the file size haven't changed since the last time
we polled. This gives applications as long as the poll interval to write
an entire log entry.