Add support for alternative encodings
This commit adds support for all encodings supported by the
`x/text/encoding` package. This can be configured with the new
`encoding` parameter of the file input.

A limitation of the current implementation is that it does not respect
the BOM in UTF16, so users will have to explicitly choose `utf-16le` or
`utf16-be`