Fix issue with default outputs in plugins
With the initial implementation of default outputs, omitting the output
for a plugin operator would cause the output variable in the templates
to be empty, defaulting to the next operator in the plugin pipeline.
This was not caught originally because in tests, only the last operator
in a plugin pipeline used the output variable. This correctly sets the
output template variable before rendering the plugin pipeline.