Utilize resources (#105)
* Utilize resources

This commit makes a series of changes to make use of the new top-level
resource field.

The contents of `$resource` use the same resource standard as Open
Telemetry. See here for details: https://github.com/open-telemetry/opentelemetry-specification/tree/master/specification/resource/semantic_conventions

This adds support for accessing fields with dots in them using syntax like
`$record['field.with.dots']`.

This also modifies the `host_metadata`, `k8s_event_input`,
and `k8s_metadata_decorator` operators to add resource identifiers to
the entry.

Additionally, it adds support in the `google_cloud_output` for using the
resource field to add logs to a monitored resource. In the process, it
moves `google_cloud_output` into its own package for better organization
of its multiple files.



Includes a fix for a minor issue where rendering plugin ID in the template would set it as <no value> if using a default plugin ID.

Includes a fix for an issue where not all k8s events had a LastTimestamp, so now we prefer EventTimestamp, but fall back to LastTimestamp, then FirstTimestamp